+++ src/main/java/com/example/k5_iot_springboot/entity/view/OrderTotalsView.java
@@ -21,7 +21,7 @@ public class OrderTotalsView {
 
     private Long user_id;
     private String order_status;
-    private Integer order_total_amount;
+    private Long order_total_amount;
     private Long order_total_qty;
 
     @Column(name = "ordered_at")
